requirements.txt
flask==2.3.2
pykakasi==2.3.0
jaconv==0.4.0
openai==1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0